eth/protocols/snap: add sync_v1.go and move heal-related types to sync_v1.go
